+++ py/src/validb-use/example.py
@@ -11,7 +11,7 @@ from validb import (
     Rule,
     validate_db,
 )
-from validb.datasources import SQLAlchemyDataSource
+from validb.datasources.sqlalchemy import SQLAlchemyDataSource
 from validb.rules.sqlalchemy import SQLAlchemyRule
 from validb.csvmapping import SimpleDetectionCsvMapping
