+++ py/src/validb/datasources/__init__.py
@@ -1,8 +1,6 @@
 from ._datasource import DataSource, DataSources
-from ._sqlalchemy import SQLAlchemyDataSource
 
 __all__ = [
     "DataSource",
     "DataSources",
-    "SQLAlchemyDataSource",
 ]
