+++ py/src/validb/datasources/sqlalchemy/__init__.py
@@ -1,0 +1,5 @@
+from ._datasource import SQLAlchemyDataSource
+
+__all__ = [
+    "SQLAlchemyDataSource",
+]
