+++ py/src/validb/rules/sqlalchemy/_rule.py
@@ -2,7 +2,8 @@ import typing as t
 
 from sqlalchemy.sql import text
 
-from ...datasources import DataSources, SQLAlchemyDataSource
+from ...datasources import DataSources
+from ...datasources.sqlalchemy import SQLAlchemyDataSource
 from ..._embedder import Embedder
 from ..._embedded_vars import EmbeddedVariables
 from ..._detected import ID, MSG, DETECTION_TYPE, Detected, DetectedType
